Moved title selector from its own file into our other file that will hold all of the markdown selectors regarding lists and moved the test accordingly
