+++ src/components/App.js
@@ -1,7 +1,6 @@
 import React from 'react';
 import Document from './markdown/Document';
 
-
 export default function App() {
   return (
     <>
