--- src/selectors/getTitlesSelector.js
@@ -1,6 +1,0 @@
-export const getTitles = state => {
-  return state.markdownListReducer.markdownList.map(markdown => {
-    return markdown.title; 
-  });
-};
-
